Backend/ml model/requirements.txt
flask==3.0.2
flask-cors==4.0.0
tensorflow==2.16.1
pillow==10.2.0
numpy==1.26.4 